requirements.txt
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.10.0
numpy>=1.24.0
scipy>=1.10.0
tqdm>=4.65.0
click>=8.1.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
auto-gptq>=0.4.0
autoawq>=0.1.0
optimum>=1.12.0
hqq>=0.1.0
safetensors>=0.3.0
vllm>=0.2.0 